pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repodanta"
version = "0.1.0"
description = "repository intelligence engine"
authors = [{name = "yogan"}]

dependencies = [
    "faiss-cpu",
    "numpy",
    "requests",
    "sentence-transformers"
]

[project.scripts]
repodanta = "repodanta.cli:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["repodanta*"]